pyproject.toml
[tool.poetry]
name = "wind_app"
version = "0.2.0"
description = ""
authors = ["Upnedra Pratap Kushwaha <upendrapratap001@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.1.2"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.6"
pytest = "^6.1.1"
celery = "4.4.1"
redis = "^3.5.3"
django-debug-toolbar = "^3.1.1"
django-extensions = "^3.0.9"
django-environ = "^0.4.5"
django-model-utils = "^4.0.0"
django-probes = "^1.5.0"
django-allauth = "^0.42.0"
boto3 = "^1.16.9"
Pillow = "^8.0.1"
pandas = "^1.1.4"
stripe = "^2.55.0"
requests = "^2.25.0"
pytest-django = "^4.1.0"
#django-celery-beat = "^2.1.0"
gunicorn = "^20.0.4"
flower = "^0.9.7"
djangorestframework = "^3.12.4"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
